Test fixes for the `BasicEditorWorkflows_LevelEntityComponentCRUD` test (#15834)
* Avoid concatenating a string when there's no room for the concatenation

* Avoid using the `ClassBuilder` after it has been destroyed

`behaviorContext->Class<LuaMaterialFunctorAPI::RuntimeContext>`
returns a `ClassBuilder<LuaMaterialFunctorAPI::RuntimeContext>` by
value. That object is is then passed to ClassBuilder::operator->(),
which transforms it into a pointer. Then it is passed to a bunch of
functions that all return a `ClassBuilder*`, which is why they can chain
together to make `ClassBuilder::Method()` calls. By the end of that
expression on line 320, there is no owner of the `ClassBuilder` value, and
its destructor runs. However, a pointer to it is saved in the builder
variable declared on line 320. That invalid pointer is then passed to
`LuaMaterialFunctorAPI::ReadMaterialPropertyValues::ReflectSubclass`, and
undefined behavior results.

This retains the object returned by `BehaviorContext::Class()` until the
end of the enclosing function.

* Fix off-by-one error when allocating XML string

This code uses `rapidxml::allocate_string` to allocate enough space for a
string. The code afterwards explicitly null-terminates the string, so it
tries to allocate enough space for the current length of the string + 1.
However, the `allocate_string` call also copies the source string into
place, copying the same number of bytes as the `size` argument. Because we
asked for one more than the current size of the string, this code ends up
in an invalid read of size 1.

This fixes that bug by doing the copy ourselves. We allocate enough space
for the null terminating byte, then copy just the source string (whose size
does not include space for the null terminator), then finally explicitly
null terminate it.

Signed-off-by: Chris Burel <burelc@amazon.com>
(cherry picked from commit 4b65521188a2a0beb399eca7b692216d2fe3d209)