Added `o3de_add_variant_dependencies_for_gem_dependencies` cmake function
The `o3de_add_variant_dependencies_for_gem_dependencies` function will
add CMake TARGET dependencies for a specified Gem and a set of Gem variants.
For each Gem variant specified a check for a TARGET of `${gem_name}.${variant}` will occur on the gem.
If such a TARGET exists, then each of the specified Gem's dependent gems
listed in its gem.json `"dependencies"` field will have the
corresponding `${dependency}.${variant}` TARGET added as a dependency.

For example in the MultiplayerSample project, there is a
[MPSGameLift](https://github.com/o3de/o3de-multiplayersample/blob/development/MPSGameLift/gem.json) gem that has a dependency on the [AWSGameLift](https://github.com/o3de/o3de-multiplayersample/blob/development/MPSGameLift/gem.json#L23) gem.

If the desire for the MPSGameLift gem when active to to have the AWSGameLift gem be
added as a load dependency automatically for the Gem Variants of
`Clients`, `Servers` and `Unified`, then the MPSGameLift gem can use the
`o3de_add_variant_dependencies_for_gem_dependencies` function to add
the following dependencies
```
MPSGameLift.Clients -> AWSGameLift.Clients
MPSGameLift.Servers -> AWSGameLift.Servers
MPSGameLift.Unified -> AWSGameLift.Unified
```

This would cause the `cmake_dependencies.multiplayersample.<launcher>.setreg` files to add both MPSGameLift shared libraries and AWSGameLift shared libraries to the list of gems to load.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>