Updated o3de CLI register command to generate a CMakePresets file for project registration
When a project is registered a
`<project-root>/user/cmake/engine/CmakePresets.json` will be
added/updated which contains an "include" entry to the registered
engine's ``CMakePresets.json` file

The Default and Minimal Project templates have had their
`<project-root>/CMakePresets.json` file to include the generated
"user/cmake/engine/CMakePresets.json" file.

This means that new projects will now opt-in to this functionality by
default.

Added a CMakePresets.json file for the AutomatedTesting project that
includes the file in the CMakePresets.json it's engine root directory.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>