Adding Archive format complete ArchiveWriter implementation
The ArchiveWriter implementation can mount an archive file, where it can
add/remove contentfiles from the archive, as well as optionally compress the
content files being added.

It also supports compressing the Archive Table of Contents using a
compression algorithm registered with the Compression gem.

The ability to write to a new archive or an existing archive is
supported by the implementation.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>