Fix deadlock in Task Systems when a TaskGraphEvent is used with 0 tasks
When the `TaskGraphEvent::Wait` function is used to wait for the
completion of task, if there are no task added to the TaskGraph, a
deadlock will occur.

This changes fixes the issue by invoking `TaskGraphEvent::Signal` if
there are no tasks on the Compiled Task Graph when submitted to the
TaskExecutor.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>