Fixed CompressContentFileAsync logic to properly compress file content
The start of each compressed block were being overridden with \0 bytes
instead of being added as padding bytes to the compressed block to reach
512-byte alignment.

Set the compression algorithm ID in the  AddFileToArchive result value
structure.
It is now properly returned as part of the result

Renamed the `ArchiveBlockSize` constant to
`ArchiveBlockSizeForCompression` to indicate that the block size is for viewing a 2 MiB chunk of an uncompressed file when sending it through a compression algorithm.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>