Fixed more ArchiveWriter functionality and added UnitTest
Added a test to validate that an uncompressed file > 6 MIB creates 2 block line entries.

Added test to validate that a file can be added and removed from the
same mounted archive instance in a single Archive Writer.

Also validated file additional and removal across multiple Archive
writen instances, where the first instance, adds a file and the next instance remove a file from the archive.

A test was added to make sure that an existing file in the Archive can be updated using the `ArchiveWriter::AddFileToArchive` function.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>