Refactor the file structure of the Archive files
Public API Files that can be used in both the Client and Tool modules of
other Gems  have been moved to the `Include/Archive/Clients` directory
Public API that can be used in only the Tools modules have been moved to
the `Include/Archive/Tools` directory.

The ArchiveToc* files have been moved to the Archive.Private.Object
target as they only need to be used as part of implementing the Archive
Reader and Writer logic and should not be part ofthe public API.
They have been relocated to the `Source/Clients` directory.

Add a new ArchiveWriter test to validate that it can write out a
compressed table of contents(TOC) to a GenericStream.
Reading of an archive with a compressed TOC was also tested.

The testing found issues in the Compression and Decompression logic of
the Table of Contents in the ArchiveWriter.cpp and those logic have been
addressed.

Updated the Archive format to write out an 8-byte magic byte sequence at the beginning of the Table of Contents section in order to detect when an invalid archive TOC of contents has been read.

This is used to avoid crashes within the Archive code and instead inform
the user of an invalid Archive table of contents.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>