Updated the SetregBuilder to generate per launcher aggregate setreg
The Setreg builder will now generate a `bootstrap.<launcher-type>.<config>.setreg` file.

There are now 9 setreg files that are generated
1. `bootstrap.client.debug.setreg` - loads in the `${project}.GameLauncher` (ClientLauncher) built in the debug configuration
1. `bootstrap.client.profile.setreg` - loads in the `${project}.GameLauncher` (ClientLauncher) built in the profile configuration
1. `bootstrap.client.release.setreg` - loads in the `${project}.GameLauncher` (ClientLauncher) built in the release configuration
1. `bootstrap.server.debug.setreg` - loads in the `${project}.ServerLauncher` (ServerLauncher) built in the debug configuration
1. `bootstrap.server.profile.setreg` - loads in the `${project}.ServerLauncher` (ServerLauncher) built in the profile configuration
1. `bootstrap.server.release.setreg` - loads in the `${project}.ServerLauncher` (ServerLauncher) built in the release configuration
1. `bootstrap.unified.debug.setreg` - loads in the `${project}.UnifiedLauncher` (UnifiedLauncher) built in the debug configuration
1. `bootstrap.unified.profile.setreg` - loads in the `${project}.UnifiedLauncher` (UnifiedLauncher) built in the profile configuration
1. `bootstrap.unified.release.setreg` - loads in the `${project}.UnifiedLauncher` (UnifiedLauncher) built in the release configuration

The Unified Launcher code has been updated to add the launcher type of
the running launcher to the Settings Registry on execution.

The AzGameFramework GameApplication::MergeSettingsToRegistry now reades the launcher type setting to determine the `bootstrap.<launcher-type>.<config>.setreg` file to load.

An additional change is that the launcher type is now added as a
specialization in the settings registry which is used when merging
folders.

For example this means when running the `${project}.GameLauncher`,
settings that will only load in the client launcher can also be placed
in a `<custom-name>.client.setreg file.
For the server launcher a `<custom-name>.server.setreg` can be used to
load server specific settings.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>