Fixed calculation of `bootstrap.<launcher-type>.<config>.setreg` SubIDs
The SubID calculation was taking into account the project name and
project launcher executable name which would cause a different Asset
SubID to be generated based on the active project.

That is not needed since each project has it's own Asset Cache and
therefore the SubIDs values should be the same across all projects.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>