Adding back the "game" specialization tag to the SettingsRegistryBuilder
Also adding a "launcher" specialization tag to the Settings Registry
Builder.

The tags are used to merge *.setreg and *.setregpatch files that
contains a sequence of `*.<tag-name>.*` as part of the filename.

This will fix the issue with the streamer.game.setreg files not being
merged into the aggregate `bootstrap.<launcher>.<config>.setreg` files.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>