Fixed for null fields being removed from every file being merged into the SettingsRegistry
The issue was that before the actual merge operation could take place,
the JSON Import logic would run the JSON Merge Patch algorithm over the
JSON data to be merged and that would remove null fields.

This would cause an issue when the desire was to actually maintain the
`null` fields in the merged JSON result.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>