Refactored duplicated logic in the Json Importer into Lambda functions
The `JsonImporter::PatchField` function now uses a `CreateObjectOrArray`
and `MergeField` lambda functions to prevent copy/paste of the logic to
copy a field value from the source JSON to the target JSON value.

Moved the `ReporterString` fixed string type alias to
JsonSerializationResult.h to reduce the need to make a string type alias
for popuplating string like structures to pass into the
`JsonIssueCallback` function when reporting on Json Serialization
operations.

Generalized the .setreg filenames in the
`SettingsRegistryTest.MergeSettingsFile_JsonWithImport_MergesImportedFiles_InOrder` unit test to not mention companies operating systems or device names.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>