Fix invalid replace arg count error handling (#16368)
* Fix invalid replace arg count error handling

When the replacement count is odd, one of the pairs will be missing resulting in a script error.  This changes checks for this case and provides better error messaging and a unit test.

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>

* Provide --force hint

Fixes #13644

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>

* Fix file uri download paths not working

When using a file:// path (missing slash) downloads in Project Manager would failed with an unknown error.  This change fixes the issue with the slashes and adds error messaging when using file: uris which are common for testing remote repos before publishing.

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>

* add replacement pairs to error log

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>

* fix checking empty path

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>

---------

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>