Updated the verify_tool function to grab the error output from the exception output variable
When validating gradle, if an error occurs it was not sent to stderr,
but instead stdout.

Therefore the error message would be loss.

With the fix all output is logged and errors are surfaced.
For example one of the issues with validating gradle is that JAVA_HOME
is not set correctly.

That kind of error message now outputs to the terminal

```
gradle cannot be resolved or there was a problem determining its version
number. Either make sure its in the system path environment or a valid
path is passed in through the --gradle-install-path argument.

ERROR: JAVA_HOME is set to an invalid directory: C:\Program
Files\Java\jdk1.8.0_311

Please set the JAVA_HOME variable in your environment to match the
location of your Java installation.
```

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>