Added SerializeContext reflection for the SettingsRegistryScriptProxy
The SettingsRegistryScriptProxy is a proxy class which exposes access to
the Settings Registry in a more user friendly way to scripters.

Because it is available in ScriptCanvas and is a data slot of a script
canvas node, it needs to be reflected in order to store the type of the
data slot into the a .scriptcanvas file.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>