Update the AttributeReader code to support reading integers from `long` and `unsigned long`.
The AttributeReader had a feature gap that attempting to read an attribute that was of type `long` and `unsigned long` wasn't possible.

This issue is more apparent on Linux where the DPE stores it's attributes in a Dom::Value where it uses the standard C type aliases  of `int64_t` and `uint64_t`

However, while the standard type aliases are always 64-bits across platforms, they are *NOT* the same underlying type.

So on Windows `int64_t` and `uint64_t` are of type `long long` and `unsigned long long` respectively.

However on Linux and Android `int64_t` and `uint64_t` are of type `long` and `unsigned long long` respectively.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>