Adding dedicated long/unsigned long and s64/u64 network serializers
This allows serialization of the `size_t` type alias through the network
serializer on MacOS/iOS.

The current issue is that on MacOS a serializer is exposed for
`uint64_t`, which on MacOS is a type alias to `unsigned long long`,
however the `size_t` type alias refers to `unsigned long`.

Now because there `unsigned long` overload on MacOS/iOS that type cannot
be serialized.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>