Adding overloads for DOM Utils for ValueFromType and MarshalOpaqueValue which can accept a void* and a type trait structure
This allows dynamic marshalling of data to a Dom::Value in a type erased manner.

Added registration of a function that can create an AZStd::any action handler which can be used to initial an AZStd::any in a type erased manner using a `void*` and `AZStd::any::type_info` object.

Updated the `PropertyEditorAPI_Internals.h` to use the `void*` overload ValueFromType overload to write the actual type stored in the PropertyHandler to a new Dom::Value.

Fixed nullptr deferenced in the AZ::Transform Row Widget.

resolves #16439

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>