Adding overloads for DOM Utils for ValueFromType and MarshalOpaqueValue which can accept a void* and a type trait structure (#16629)
* Adding overloads for DOM Utils for ValueFromType and MarshalOpaqueValue which can accept a void* and a type trait structure

This allows dynamic marshalling of data to a Dom::Value in a type erased manner.

Added registration of a function that can create an AZStd::any action handler which can be used to initial an AZStd::any in a type erased manner using a `void*` and `AZStd::any::type_info` object.

Updated the `PropertyEditorAPI_Internals.h` to use the `void*` overload ValueFromType overload to write the actual type stored in the PropertyHandler to a new Dom::Value.

Fixed nullptr deferenced in the AZ::Transform Row Widget.

resolves #16439

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

* Added a native visualizer for Dom::Path that allows viewing the full
path as a single string.

It supports up to 48 path segments (/1/2/foo/bar/.../baz) which should
be more than enough to visualize any real Dom path in the debugger.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

* Fixed Document Property Editor removal of Assets using the UI

The Asset Json Serializer would not update the object being loaded if
the AssetId stored in JSON was the invalid Asset Id of
"id={00000000-0000-0000-0000-000000000000}:0"

The Asset Manager has been updated to not attempt to Create or Find
Assets using the Invalid asset Id of Uuid "0".

Finally the ReflectionAdapter VisitObjectBegin method has been updated
with `AZ_Error` messages to log when copying a new value into the
reflected instance address fails in JSON Serialization.

fixes #16231

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

* Adding trace message suppression to the Slice Entity Ownership Test.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

* Fixed nullptr dereference in the Slice Entity ownership test

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

---------

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>
Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>