Updated the AssetJsonSerializer to preserve the Asset Id on load
There are two scenarios where the AssetId wasn't be preserved with the
changes in commit 3393f127db73e410449a06e5daa417786c06c4c2

The first scenario is if the Asset Id is not registered with the loaded Asset
Catalog in the Asset Builder or if the Asset wasn't processed yet.

The second scenario is if the Asset Id is registered with the Asset
Catalog, but the Asset Type isn't registered with the Asset Manager,
therefore not allowing the Asset to be loaded successfully using it's
Autoload behavior.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>