DOM ValueFromType and ValueToType json serialization support
The DomUtils.h/.cpp `ValueFromType` and `ValueToType` functions have
been updated to use the JSON Serialization system when possible to
convert an `AZ::Dom::Value` to/from a C++ type.

This is the preferred way to store a aggregate type into
`AZ::Dom::Value` as it maintains the C++/C object structure in a Dom
Object/Array structure.

If the type doesn't support JSON Serialization, then it fallback to
being stored using an `AZStd::any` inside of the Dom::Value as an opaque
type.
The limitations of using an opaque type, is that there is no support for
comparing two `AZStd::any` objects to each other.
The only way for comparison, is the compare the address where the
`AZStd::any` objects are stored.

Due to that, the Dom::Value::operator= and AZ::DeepCompareIsEqual can
only compare the internal `AZStd::shared_ptr<AZStd::any>` instance
memory address to compare if two opaque types are equal and not a deep
compare of the fields of the stored aggregate type.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>