Modified the ReflectionAdapter `VisitObjectBegin` function `StoreValueIntoPointer` lambda to account for marshaling Dom ObjectTypes
The `StoreValueIntoPointer` value was only accounting for visiting
pointers to C++ types through the TryMarshalValueToPointer function
call.

Now that the Dom Utils `ValueFromType` function has been updated to
write C++ structs using Json Serialization to the Dom::Value when
possible, the scenario where a Dom::Value which is an "Object" is being
visited.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>