Moved `TryMarshalValueToPointer` logic earlier in the ExtractOpaqueValue lambda
The flow now attempts to check if the Dom::Value is a Dom::Object which
has a pointer field that can be deserialized first.
If that fails, then the JSON Serialization system is used to load Dom
Value object into a C++ type.

This is needed as the AssetJsonSerializer needs to suceed in its
deserialization operation, when all the relavant asset fields are
missing such as the "assetId" field.

The result of the operation indicates that JSON Serialization
has used the default value to intialize the Asset.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>