Associative Container low-level API improvements
Updated the IAssociativeDataContainer interface to expose a
`GetAssociativeType` function that can return whether the associative
container is a set, map, unordered_set or unordered_map

Updated the IDataContainer interface to expose a `IsSequenceType`
function which returns if the DataContainer is a sequence type such as a
(Fixed)Vector, (Forward)List or Array.

A const member overload for `GetAssociativeContainerInterface` function have added to allow calling on a const IDataContainer pointer.

Two new console commands have been added to the Edit Context which can
log the name of fields that have been reflected to the Edit Context that
are sequence types or associative types.

The console commands are `ed_logReflectedAssociativeContainerFields` and `ed_logReflectedSequenceContainerFields`.

The commands output the name of the class in which the field was
reflected on and the name of the field has specified to the EditContext
`DataElement` function.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>