Modifications to LegacyReflectionBridge associative container support
The mapped value of an map or unordered_map being visited by the
LegacyReflection Bridge now has a "KeyValue" attribute added to it,
which stores off the address of the key value and the type ID of the key
value.

This will allow the ReflectionAdapter to query the Key value data when
building the Adapter for the Mapped value data.

Also added support to get the string name of an Enum reflected to the SerializeContext.
There is no need to reflect an Enum to the Edit Context to read its
string name anymore.

Updated the `void*` members for the `StackEntry` struct in the
LegacyReflectionBridge to be `AZ::PointerObject` members which supports
storing an AZ::TypeId as well

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>