Fixing the gradle_windows.cmd script to set the JAVA_HOME env (#16942)
* Fixing the gradle_windows.cmd script to set the JAVA_HOME env

Previously the gradle_windows.cmd script would set the JDK_HOME
environment variable which is not used by the Android SDK Manager and
Gradle.

This worked on Jenkins due to the system JAVA_HOME variable being set to
Java 8 in the older AMI images for building Android on Jenkins, but once
the JAVA_HOME environment variable was set to a version other than 8 in
the newer AMIs,
the Jenkins build is now failing due to SDK Manager trying to use that
version.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

* Update the gradle_windows.cmd Java check to check the Windows registry before the environment

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

---------

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>
Signed-off-by: Steve Pham <82231385+spham-amzn@users.noreply.github.com>