Updating references to the AllocationsRecord::Mode enum
The references now explicit mention the "Mode" enum name itself.
This is needed for the next commit where the "Mode" member of the
AllocationsRecord class is being changed to a type alias for an
"AllocatorRecordMode" enum declared using AZ_ENUM to allow AZ Console
system to be able to convert an string specifying one of the `AllocationRecordMode` enum options to the actual enum type.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>