Updated the ComponentApplication to support initalizing tracking of the allocator through a startup config file
The Component Application can now initialize the allocations settings using the following algorithm.

First it checks if there is an O3DE_STARTUP_CFG_FILE environment variable set.
If that is set that uses used as the startup config file

Next if the O3DE_STARTUP_CFG_FILE isn't set, a file of `<executable-directory>/Registry/startup.cfg` is checked for the startup config file
If the `<executable-directory>/Registry/startup.cfg` doesn't exist, the path of `~/.o3de/Registry/startup.cfg` is used for the startup config file.

Finally the startup config file location can be overridden with a new command line option of `--startup-cfg-file`
The command line option is parsed with using no heap allocations, before allocators are created.

If the startup config file is a valid file, it is opening up using the ConfigParser and checked for keys of the form "allocator_tracking_<allocator-name>"
The "allocator_tracking_<allocator-name>" is used to specifying the recording mode of the allocation records that should be used for the allocator with name "<allocator-name>"

For example given a cfg file with the following content
```
allocator_tracking_SystemAllocator = true
allocator_tracking_OSAllocator = RECORD_STACK_IF_NO_FILE_LINE
allocator_tracking_PoolAllocator = false
```

A mapping of allocator name to the AllocationRecordMode enum is registered with the AZ::AllocatorManager

1. The SystemAllocator has allocation tracking turned ON using the recording mode of `RECORD_FULL`
2. The OSAllocator has allocation traking turned ON using the recording mode of `RECORD_STACK_IF_NO_FILE_LINE`
3. The PoolAllocator has allocations tracking turned OFF and its recording mode set to `RECORD_NO_RECORDS`

The "allocator_tracking_<allocator-name>" setting can accept either a bool value or one of the options of the AllocationRecordMode located in AllocationsRecords.has
If a bool option is used, then the value of `true` maps to the AllocationRecordMode of `RECORD_FULL` and `false` maps to `RECORD_NO_RECORDS`

Two new Console commands of `sys_DumpAllocationRecordsSlow` and `sys_DumpAllocationRecordsInRange` have been added which can dump the allocation records of all allocators or a user specified allocator

The `sys_DumpAllocationRecordsSlow` command dumps ALL allocations of any specified allocator or all allocators if no allocator name is provided.
This function can take a long time to run (on the order of 10s of minutes) if dumping the allocation records of the SystemAllocator

The `sys_DumpAllocationRecordsInRange` command forces the user to specify a range of indices to cap how many allocations should be dumped from any following allocators if they are specified to the command or ALL allocators if not.
This is the better function to use if the allocator has a large number of records (on the order of 100000's)

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>