Update the AllocatorBase class to support specifying the flag to enable profiling on construction
The AllocatorBase also swaps the order of logic in PostCreate/PreDestroy.
The AllocationsRecords structure is now created before registering the allocator with the AllocatorManager and destroyed after unregistering the allocator with the AllocatorManager
This is needed as the AllocatorManager now checks a map of allocator name -> allocation tracking structure in its RegistryAllocator function to determine the record tracking mode for allocator

The ChildAllocatorSchema has been updated to inherit from the AllocatorBase class to allow it to automatically register with the AllocatorManager on creation.
It has also been provided AZ_RTTI information which allows derived allocator classes to override the allocator name via the RTTI system.

On a related change, the SimpleSchemaAllocator, now unregisters the schema it is managing with the AllocatorManager, as the SimpleSchemaAllocator itself registers with the AllocatorManager.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>