Updated the NoWindow constant to be an empty string.
The Trace `Output`, `RawOutput` and `Printf` will not default to the
NoWindow constant if a nullptr window value is supplied via a `const
char*`.

In addition Trace fnuction overloads which accept a `std::nullptr_t`
type for the window parameter have been added to cause a
compile error when `nullptr` has been explicitly passed for the window
parameter.
The window parameter is expected to be valid c-string value.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>