Updated the AZStd container calls to allocate to cast the call to allocate to void pointer first using static_cast before a reinterpret_cast to the desired pointer type.
This change is in anticipation of the next couple of commits to update
the AZ IAllocator api allocate function to return struct encapsulating
the allocated pointer and the amount of bytes allocated instead of just
a void*.

The new struct of `AllocateAddress` supports an implicit conversion
t void* or an explicit conversion to any pointer type.
However to trigger the conversion operator a static_cast c-style cast
must be used. A reinterpret_cast will not use a class conversion
operator.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>