Updated the AZ Allocator API
The `IAllocator::allocate` and `IAllocator::reallocate functions now returns a new struct called
AllocateAddress which is implicitly convertible to a void*.
The AllocateAddress struct encapsulates the allocated memory address
along with the actual amount of memory allocated by the underlying
allocator.

The `IAllocator::deallocate` function now returns a `size_t` containing
the amount of memory that was just de-allocated at the specified
address.

The ChildAllocatorSchema and SimpleAllocatorSchema classes have been
updated to use the new return values to track in an atomic integer the
amount of memory that was allocated when making calls through it's
interface.

It was returning the amount of allocated memory from the underlying
allocator, which is not very useful for diagnostic purposes.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>