Added a `O3DE_STARTUP_CFG_FILE_CHECK_OVERRIDE` CMake Cache variable
The `O3DE_STARTUP_CFG_FILE_CHECK_OVERRIDE` allows the user to override
whether they to allow checking for a startup config file for
initializing allocator tracking when a ComponentApplication is
initialized in O3DE.

The `O3DE_STARTUP_CFG_FILE_CHECK_OVERRIDE` is unset by default.
In that case the non-development build configurations of debug/profile
support checking for a startup.cfg on launch, and the release
configuration does not.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>