Added a new `AZ_CHILD_ALLOCATOR_WITH_NAME` macro
The `AZ_CHILD_ALLOCATOR_WITH_NAME` macro fills out the boilerplate
associated with creating an AZ Allocator that is derived from the ChildAllocatorSchema.

The macro creates an allocator with the the "_ALLOCATOR_CLASS"
parameter to the macro as an allocator derived from the
ChildAllocatorSchemea and an AZStdAlloc alias for an AZStd compatible allocator made by concatenating the
"_ALLOCATOR_CLASS" parameter with the literal token of "_for_std_t".

Therfore `AZ_CHILD_ALLOCATOR_WITH_NAME(FooAllocator, ...)` would result in a type
identifier of `FooAllocator` and `FooAllocator_for_std_t` to be
available.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>