Clear the global functions that the Lua Material functor calls
When the `ScriptContext::Execute` function loads and executes a Lua
script it adds any `functions` at script scope to the Lua global table.

The function that are added however are not remove from the global table
on subsequent invocations on `ScriptContext::Execute`, so if a new
script was loaded that didn't replace existing functions in the global
table, then attempting to invoke the existing function would call a
stale version of the function instead of not finding the function and
properly failing.

This was causing newer loaded lua scripts invoked in the `LuaMaterialFunctorSourceData` to call older scripts version of the `GetShaderOptionDependencies` function when the newer script didn't actually implement the function.

Therefore incorrect shared option names was being returned for the newer
script where it should have returned zero shader options.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>